Fix theme and GlobalStyle imports due files relocation
