+++ src/main.tsx
@@ -4,9 +4,9 @@ import { Provider } from "react-redux";
 import App from "./components/App/App";
 import "./index.css";
 import { store } from "./store";
-import GlobalStyle from "./styles/GlobalStyle";
+import GlobalStyle from "./styles/GlobalStyle/GlobalStyle";
 import { ThemeProvider } from "styled-components";
-import theme from "./styles/theme";
+import theme from "./styles/theme/theme";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
